backend/requirements.txt
# Core Django
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt
asgiref==3.8.1
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.6
redis==5.0.1
amqp==5.3.1
kombu==5.5.3

# API Clients
openai>=1.57.0
google-generativeai==0.3.2
praw==7.8.1
tweepy==4.14.0

# HTTP and Async
aiohttp==3.9.3
requests==2.32.3
urllib3==2.4.0

# Analysis Tools
vaderSentiment==3.3.2
emoji==2.10.1
numpy==1.26.4

# Image Processing
Pillow==10.2.0

# Environment and Utils
python-dotenv==1.0.1
pathlib==1.0.1
gunicorn==21.2.0
huggingface-hub>=0.20.3
safetensors>=0.4.1

# Deep Learning and GPU Support
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.41.0
sentencepiece>=0.1.99
accelerate>=0.27.2
bitsandbytes>=0.41.1
einops>=0.7.0

# Additional dependencies
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34
nltk==3.8.1
protobuf>=4.25.2

# Testing
pytest
pytest-django

sentry-sdk

python-json-logger

django-debug-toolbar

drf-spectacular[sidecar]
